Parse simulation settings from YAML
This makes it easier to run simulations without adapting Python code.
Additionally, it makes it easier to adapt simulation settings at runtime
(e.g., for convergence studies).
Simulation parameters are stored in a dataclass Params.
This class is properly parsed from a YAML file and used by the solvers.
This also simplifies the constructors of the solvers.
Related, smaller changes:
1. Users should now call coupled.py from top-level directory
2. Move Enum for boundary conditions to setup_simulation.py
3. Remove some __name__ == "__main__" guards for scripts that
should not be executed.
4. Update README with new dependencies/instructions.